pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohammad Hossein Rostamdar",email = "m.hosein.rostamdar@gmail.com"}
]

requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "schedule (>=1.2.0,<2.0.0)",
    "click (>=8.0.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
